Upgraded Table authoring to Angular. #2902
