+++ src/main/webapp/site/src/app/teacher-hybrid-angular.module.ts
@@ -52,6 +52,7 @@ import { ConceptMapAuthoring } from '../../../wise5/components/conceptMap/concep
 import { DrawAuthoring } from '../../../wise5/components/draw/draw-authoring/draw-authoring.component';
 import { MatchAuthoring } from '../../../wise5/components/match/match-authoring/match-authoring.component';
 import { LabelAuthoring } from '../../../wise5/components/label/label-authoring/label-authoring.component';
+import { TableAuthoring } from '../../../wise5/components/table/table-authoring/table-authoring.component';
 
 @NgModule({
   declarations: [
@@ -87,6 +88,7 @@ import { LabelAuthoring } from '../../../wise5/components/label/label-authoring/
     RubricAuthoringComponent,
     StatusIconComponent,
     StepInfoComponent,
+    TableAuthoring,
     WorkgroupInfoComponent,
     WorkgroupNodeScoreComponent,
     WorkgroupSelectAutocompleteComponent,
