+++ src/main/webapp/wise5/components/table/tableAuthoringComponentModule.ts
@@ -2,14 +2,17 @@
 
 import * as angular from 'angular';
 import { TableService } from './tableService';
-import { downgradeInjectable } from '@angular/upgrade/static';
-import TableAuthoring from './tableAuthoring';
+import { downgradeComponent, downgradeInjectable } from '@angular/upgrade/static';
 import { EditTableAdvancedComponent } from './edit-table-advanced/edit-table-advanced.component';
+import { TableAuthoring } from './table-authoring/table-authoring.component';
 
 const tableAuthoringComponentModule = angular
   .module('tableAuthoringComponentModule', ['pascalprecht.translate'])
   .service('TableService', downgradeInjectable(TableService))
-  .component('tableAuthoring', TableAuthoring)
+  .directive(
+    'tableAuthoring',
+    downgradeComponent({ component: TableAuthoring }) as angular.IDirectiveFactory
+  )
   .component('editTableAdvanced', EditTableAdvancedComponent)
   .config([
     '$translatePartialLoaderProvider',
